test(signals): move test to signal-state file
